Create a streamer for calculating the population and sample variance

Variance creates a streaming algorithm that can be used to calculate the population and sample variance of incoming values

Format

An R6Class generator object

Methods

Public methods

Variance$new()

Variance$value()

Variance$update()

Variance$clone()

Method new()

Creates a new Variance streamer object.

Usage

Variance$new(x = NULL)

Arguments

x

values to be used during initialization (optional)

Returns

The new Variance (invisibly)

Examples

variance <- Variance$new()

Method value()

Returns the current value of the Variance.

Usage

Variance$value(sample = FALSE)

Arguments

sample

for choosing between the population and sample variance. If `TRUE` the sample variance is returned. If `FALSE` the population variance is returned.

Returns

The updated Variance (invisibly)

Examples

variance <- Variance$new(x = c(1,2,3,4))
variance$value()
#> [1] 1.25
variance$value(sample = TRUE)
#> [1] 1.666667

Method update()

Updates the Variance with a stream of new values

Usage

Variance$update(x)

Arguments

x

a vector of values to update the Variance

Returns

The updated Variance (invisibly)

Examples

variance <- Variance$new(c(1,2))
variance$update(c(3,4))

Method clone()

The objects of this class are cloneable with this method.

Usage

Variance$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

variance <- Variance$new(c(1, 2))
variance$update(c(3, 4))
variance$value

#> function (sample = FALSE) 
#> {
#>     if (private$count == 0L) {
#>         return(0)
#>     }
#>     if (sample) {
#>         n <- private$count
#>         return(as.numeric(private$variance * n/(n - 1)))
#>     }
#>     as.numeric(private$variance)
#> }
#> <environment: 0x7ff1e961ea08>

#> [1] 1.25

## ------------------------------------------------
## Method `Variance$new`
## ------------------------------------------------

variance <- Variance$new()

## ------------------------------------------------
## Method `Variance$value`
## ------------------------------------------------

variance <- Variance$new(x = c(1,2,3,4))
variance$value()

#> [1] 1.25

#> [1] 1.25
variance$value(sample = TRUE)

#> [1] 1.666667

#> [1] 1.666667

## ------------------------------------------------
## Method `Variance$update`
## ------------------------------------------------

variance <- Variance$new(c(1,2))
variance$update(c(3,4))